New EPFL Courses search engine plugin (CDH-SHS)
A plugin developed for the CDH-SHS section to be able to perform search on the section courses using many cumulative criterias.
The plugin contain a wp-admin section for the settings and to manage data update using ISA webservices or CSV import. Please enter the commit message for your changes. Lines starting